backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.18
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Speech-to-text (optional - install separately if needed)
# openai-whisper==20231117
# torch>=2.0.0

# Language detection and processing
langdetect>=1.0.9

# Gemini API (for summarization/translation only)
google-generativeai==0.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx>=0.26.0

# File uploads
cloudinary>=1.36.0
